--- a/deps/Versions.make
+++ b/deps/Versions.make
@@ -15,8 +15,7 @@ CSL_JLL_NAME := CompilerSupportLibraries
 
 # Clang (paired with LLVM, only here as a JLL download)
 CLANG_JLL_NAME := Clang
-# equivalent to libLLVM 12.0.1+4
-CLANG_JLL_VER  := 12.0.1+3
+CLANG_JLL_VER  := 13.0.0+3
 
 # DSFMT
 DSFMT_VER := 2.2.4
@@ -45,15 +44,14 @@ LIBUV_VER := 2
 LIBUV_JLL_NAME := LibUV
 
 # LLVM
-LLVM_VER := 12.0.1
-LLVM_ASSERT_JLL_VER := 12.0.1+4
+LLVM_VER := 13.0.0
+LLVM_ASSERT_JLL_VER := 13.0.0+3
 LLVM_JLL_NAME := libLLVM
 
 # LLVM_tools (downloads LLVM_jll to get things like `lit` and `opt`)
 LLVM_TOOLS_JLL_NAME := LLVM
-# equivalent to libLLVM 12.0.1+4
-LLVM_TOOLS_JLL_VER := 12.0.1+3
-LLVM_TOOLS_ASSERT_JLL_VER := 12.0.1+3
+LLVM_TOOLS_JLL_VER := 13.0.0+3
+LLVM_TOOLS_ASSERT_JLL_VER := 13.0.0+3
 
 # LLVM libunwind
 LLVMUNWIND_VER := 12.0.1