--- a/deps/llvm.version
+++ b/deps/llvm.version
@@ -1,2 +1,2 @@
-LLVM_BRANCH=julia-12.0.1-4
-LLVM_SHA1=julia-12.0.1-4
+LLVM_BRANCH=julia-13.0.0-0
+LLVM_SHA1=julia-13.0.0-0