--- a/doc/src/devdocs/llvm.md
+++ b/doc/src/devdocs/llvm.md
@@ -42,7 +42,7 @@ The default version of LLVM is specified in `deps/Versions.make`. You can overri
 a file called `Make.user` in the top-level directory and adding a line to it such as:
 
 ```
-LLVM_VER = 12.0.1
+LLVM_VER = 13.0.0
 ```
 
 Besides the LLVM release numerals, you can also use `DEPS_GIT = llvm` in combination with