@nanosoldier 2021-12-09
[Your package evaluation job](https://github.com/JuliaLang/julia/pull/42602#issuecomment-988980833) has completed - possible new issues were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/pkgeval/by_hash/d24d243_vs_8197c41/report.md).