@nanosoldier 2021-12-28
Something went wrong when running [your job](https://github.com/JuliaLang/julia/pull/42602#issuecomment-1001744567):
```
NanosoldierError: error when preparing/pushing to report repo: failed process: Process(setenv(`git push`; dir="/data/nanosoldier/workdir/NanosoldierReports"), ProcessExited(1)) [1]

```
Unfortunately, the logs could not be uploaded.