@vtjnash 2021-12-28
https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/15db64c_vs_c790938/report.md Looks okay. Sorry that this new AWS machine seems to be unreliable. Hopefully we can move back to dedicated hardware soon.

@nanosoldier `runbenchmarks("collections" || "shootout" || "dates" || "string", vs=":master")`