@nanosoldier 2021-12-28
[Your benchmark job](https://github.com/JuliaLang/julia/pull/42602#issuecomment-1001867036) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/15db64c_vs_c790938/report.md).