@vchuravy 2022-01-05
Non-scientific build benchmark:

### Before:
```
Sysimage built. Summary:
Total ───────  52.908324 seconds 
Base: ───────  21.431235 seconds 40.5064%
Stdlibs: ────  31.475813 seconds 59.4912%
    JULIA /home/vchuravy/builds/julia/usr/lib/julia/sys-o.a
Generating REPL precompile statements... 36/36
Executing precompile statements... 1309/1340
Precompilation complete. Summary:
Total ─────── 109.812820 seconds
Generation ──  82.326156 seconds 74.9695%
Execution ───  27.486664 seconds 25.0305%
    LINK /home/vchuravy/builds/julia/usr/lib/julia/sys.so
```

### After:
```
Sysimage built. Summary:
Total ───────  52.104447 seconds 
Base: ───────  20.850621 seconds 40.017%
Stdlibs: ────  31.252741 seconds 59.9809%
    JULIA /home/vchuravy/builds/julia/usr/lib/julia/sys-o.a
Generating REPL precompile statements... 36/36
Executing precompile statements... 1309/1340
Precompilation complete. Summary:
Total ─────── 102.547306 seconds
Generation ──  75.177300 seconds 73.3099%
Execution ───  27.370006 seconds 26.6901%
```